+++ src/impl/particles_impl_rcyc.ipp
@@ -71,7 +71,7 @@ namespace libcloudphxx
 #else
         static_assert(sizeof(thrust_size_t) == sizeof(n_t), "");
 #endif
-        thrust_device::vector<thrust_size_t> &tmp(sorted_ijk);
+        thrust_device::vector<thrust_size_t> &tmp(sorted_ijk.get());
         thrust::copy(n.begin(), n.end(), tmp.begin());
       
         thrust::sort_by_key(
@@ -98,12 +98,12 @@ namespace libcloudphxx
 
       // for each property... 
       for(auto vec: distmem_real_vctrs)
-        detail::copy_prop<real_t>(vec->begin(), sorted_id, n_flagged); 
+        detail::copy_prop<real_t>(vec->begin(), sorted_id.get(), n_flagged); 
 
       // ... chemical properties only if chem enabled
       if (opts_init.chem_switch){
         for (int i = 0; i < chem_all; ++i)
-          detail::copy_prop<real_t>(chem_bgn[i], sorted_id, n_flagged);
+          detail::copy_prop<real_t>(chem_bgn[i], sorted_id.get(), n_flagged);
       }
 
       {
