+++ src/particles.tpp
@@ -29,6 +29,7 @@
 #include "detail/functors_host.hpp"
 #include "detail/ran_with_mpi.hpp"
 #include "detail/grid_refinement_helpers.hpp"
+#include "detail/debug.hpp"
 
 //kernel definitions
 #include "detail/kernel_definitions/hall_efficiencies.hpp"
